requirements.txt
 streamlit 
 pandas 
 seaborn 
matplotlib.pyplot
 plotly.express 
sklearn.preprocessing
LabelEncoder